Fix issue with using find_template for mustache checks
There seems to be a bug in mainline ckan where find_template for some reason converts the passed template name into bytes and then does an os.path.join with a string. This breaks. To avoid this, we'll just do essentially what it does but more specific to what we want to do (i.e. check the existence of a template, not actually resolve the path to it).