Fix race condition in map view
The map view needs window.parent.ckan.views to be available so that it can grab the filters that are in place. views can be created after the map view attempts to get it, therefore we need to wait for it to show up. This code is a bit nasty but it does the job well.